airflow/requirements.txt
apache-airflow-providers-postgres==6.1.3
affine==2.4.0
attrs==25.3.0
bcdata==0.15.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
connectorx==0.4.3
dotenv==0.9.9
GeoAlchemy2==0.17.1
geopandas==1.0.1
greenlet==3.2.2
idna==3.10
lxml==5.4.0
numpy==2.2.6
OWSLib==0.33.0
packaging==25.0
pandas==2.2.3
pendulum==3.1.0
polars-lts-cpu==1.30.0
polars-st==0.1.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.3
shapely==2.1.1
six==1.17.0
SQLAlchemy==1.4.54
stamina==25.1.0
tenacity==9.1.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xmltodict==0.14.2
